moved styled code into component.styled.ts
